requirements.txt
pytorch
hydroeval
numpy
pandas
